extracted charts to separate component files
